+++ tests/Feature/EnsureFrontendRequestsAreStatefulTest.php
@@ -5,11 +5,14 @@ namespace Laravel\Sanctum\Tests;
 use Illuminate\Http\Request;
 use Laravel\Sanctum\Http\Middleware\EnsureFrontendRequestsAreStateful;
 use Laravel\Sanctum\SanctumServiceProvider;
+use Orchestra\Testbench\Concerns\WithWorkbench;
 use Orchestra\Testbench\TestCase;
 
 class EnsureFrontendRequestsAreStatefulTest extends TestCase
 {
-    protected function getEnvironmentSetUp($app)
+    use WithWorkbench;
+
+    protected function defineEnvironment($app)
     {
         $app['config']->set('sanctum.stateful', ['test.com', '*.test.com']);
     }
@@ -72,10 +75,5 @@ class EnsureFrontendRequestsAreStatefulTest extends TestCase
         $request = Request::create('/');
 
         $this->assertFalse(EnsureFrontendRequestsAreStateful::fromFrontend($request));
-    }
-
-    protected function getPackageProviders($app)
-    {
-        return [SanctumServiceProvider::class];
     }
 }
