+++ tests/Unit/TransientTokenTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Laravel\Sanctum\Tests;
+namespace Laravel\Sanctum\Tests\Unit;
 
 use Laravel\Sanctum\TransientToken;
 use PHPUnit\Framework\TestCase;
